move 'complement' tests to separate file
