+++ types/ramda/tsconfig.json
@@ -42,6 +42,7 @@
         "test/chain-tests.ts",
         "test/clamp-tests.ts",
         "test/clone-tests.ts",
-        "test/comparator-tests.ts"
+        "test/comparator-tests.ts",
+        "test/complement-tests.ts"
     ]
 }
